- crash fix in chats; - static buffer returned, now with buffer overrun control; - warnings fixed
git-svn-id: http://svn.miranda-ng.org/main/trunk@13604 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/MSN/src/msn_auth.cpp
@@ -571,7 +571,8 @@ static int CopyCookies(NETLIBHTTPREQUEST *nlhrReply, NETLIBHTTPHEADER *hdr)
 		if (hdr) {
 			if (*hdr->szValue) strcat (hdr->szValue, "; ");
 			strcat (hdr->szValue, nlhrReply->headers[i].szValue);
-		} else nSize += strlen(nlhrReply->headers[i].szValue) + 2;
+		}
+		else nSize += (int)strlen(nlhrReply->headers[i].szValue) + 2;
 	}
 	return nSize;
 }
